requirements.txt
tensorflow==2.17.0
streamlit>=1.28.0
pillow==10.4.0
numpy==1.26.4
requests==2.32.3
scipy==1.11.4
scikit-learn==1.3.2
matplotlib==3.8.2